% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_MyFirstPackage.R
\name{as_MyFirstPackage}
\alias{as_MyFirstPackage}
\title{Creates a MyFirstPackage object from a table file}
\usage{
as_MyFirstPackage(fname, row.names = 1, header = TRUE, sep = "\\t")
}
\arguments{
\item{fname}{the name of the table}

\item{row.names}{see read.delim documentation defaults to 1}

\item{header}{see read.delim documentation defaults to TRUE}

\item{sep}{see read.delim documentation defaults to '\\t'}
}
\description{
Creates a MyFirstPackage object from a table file
}
