% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTSNE.R
\name{CalcTSNE}
\alias{CalcTSNE}
\title{Calculate the mds for the scells object}
\usage{
CalcTSNE(x, ndim = 3, type = "Rtsne", ...)
}
\arguments{
\item{x}{the scell object}

\item{ndim}{the requested dimensions in the mds (default 3)}

\item{type}{the mds type to run ('Rtsne')}

\item{...}{additional options for the mds function(s)}
}
\value{
the scell object
}
\description{
Calculate the mds for the scells object
}
