% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHeatmap.R
\name{MakeHeatmap}
\alias{MakeHeatmap}
\title{Create a heatmap using pheatmap and returning the plot object}
\usage{
MakeHeatmap(x, genes, ...)
}
\arguments{
\item{x}{the MyFirstPackage object}

\item{genes}{a gene name vector}

\item{...}{other options fead to the pheatmap function}
}
\description{
Create a heatmap using pheatmap and returning the plot object
}
